README.md: experiment (yet more) with embedding the rendered PDF
Experiments to directly embed the PDF in the README.md (so it
appears on the GitHub web UI inline with the text content) have not
yet been successful; it is not yet clear whether or not that should
be expected to work. What /is/ documented to work, however, is
embedding an image, so that's what we're doing here. I used the GIMP
to create a PNG export of the PDF, and it is the PNG that this
change references from the README.md file.