doc: README.md: add 'raw=true' param to URL for rendered PDF
Update the relative Markdown URL to the in-tree "pre-rendered" PDF
file from:

    ./rendered/ads-latex-greek-alphabet.pdf

to:

    ./rendered/ads-latex-greek-alphabet.pdf?raw=true

Without the 'raw=true' parameter, GitHub will wrap the displayed PDF
in within the GitHub web UI, which is not what we want. We want the
user to simply access the PDF directly; the 'raw=true' allows for
that (at the cost of dirtying the URL).

Furthermore, testing with Debian's 'firefox-esr' package (version
115.11.0esr-1~deb11u1, from 2024-05-15) revealed that the
embed-within-GitHub-UI mechanism is not working. Rather than showing
the PDF in the designated area of the screen, we instead see a red
box with the error message:
<quote>
    Error rendering embedded code

    Invalid PDF
</quote>

That error seems to be browser-specific, though, as 'chromium'
(version 120.0.6099.224-1~deb11u1, from 2024-01-16) did work as
expected: it renedered the PDF within the the GitHub web UI.

The above change fixes the issue on Firefox, but does not "just
work" on chromium (120.0.6099.224-1~deb11u1). In chromium, instead
of seeing the PDF presented within the browser, we are presented
with the dialog box to save the file to disk (even when we have the
browser configured to show PDFs in the browser:

    'Settings'
      -> 'Privacy and security'
           -> 'Site settings'
                -> 'Additional content settings'
                     -> 'PDF documents':
                            ( ) "Download PDFs"
                            (o) "Open PDFs in Chromium"
)

A little debugging reveals that the browser is being
redirected (HTTP 302 "Found") twice, and the final document has:

    Content-Type: application/octet-stream

Orig URL: https://github.com/salewski/ads-latex-greek-alphabet/blob/ads/2024-06-10-doc-readme-link-to-raw-pdf/rendered/ads-latex-greek-alphabet.pdf?raw=true
 redir 1: https://github.com/salewski/ads-latex-greek-alphabet/raw/ads/2024-06-10-doc-readme-link-to-raw-pdf/rendered/ads-latex-greek-alphabet.pdf
 redir 2: https://raw.githubusercontent.com/salewski/ads-latex-greek-alphabet/ads/2024-06-10-doc-readme-link-to-raw-pdf/rendered/ads-latex-greek-alphabet.pdf

The browser does not seem to be driving off of the name of the
requested file (good), since starting from the second URL
above (which ends with '.pdf') yields the same result.

So the question is, how do we get GitHub to send the RFC 3778
compliant:

    Content-Type: application/pdf
?