doc: README.md: reflow (a.k.a. "fill") paragraphs
No content changes. Just invoke 'fill-paragraph' within Emacs on the
paragraphs edited earlier today.

This keeps the raw Markdown text readable, but also allowed the
earlier changes to be made in a more "diff friendly" way (that is,
without changing the line wrapping within the paragraphs that were
being modified).